moved all elements of gameLoop into another file
