+++ ExpansionOfDomain/ExpansionOfDomain.vcxproj
@@ -132,6 +132,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="src\gameLoop.cpp" />
     <ClCompile Include="src\KeyboardControls.cpp" />
     <ClCompile Include="src\main.cpp" />
     <ClCompile Include="src\Player.cpp" />
@@ -142,6 +143,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\Controls.h" />
+    <ClInclude Include="src\gameLoop.h" />
     <ClInclude Include="src\KeyboardControls.h" />
     <ClInclude Include="src\Player.h" />
     <ClInclude Include="src\renderer.h" />
